# Conflicts:
#	Sources/Actions/Admin/Themes.php
#	Sources/Actions/Display.php
#	Sources/Actions/Post.php
#	Sources/Editor.php
#	Sources/Theme.php
#	Themes/default/Calendar.template.php
#	Themes/default/Display.template.php
#	Themes/default/GenericControls.template.php
#	Themes/default/ManageNews.template.php
#	Themes/default/ManagePaid.template.php
#	Themes/default/MessageIndex.template.php
#	Themes/default/PersonalMessage.template.php
#	Themes/default/Post.template.php
#	Themes/default/Register.template.php
#	Themes/default/Themes.template.php
#	Themes/default/css/index.css
#	Themes/default/css/jquery.sceditor.css
#	Themes/default/css/jquery.sceditor.default.css
#	Themes/default/index.template.php
#	Themes/default/scripts/drafts.js
#	Themes/default/scripts/sceditor.plugins.autolinker.js
#	Themes/default/scripts/sceditor.plugins.smf.js
#	Themes/default/scripts/script.js
#	Themes/default/scripts/theme.js
#	Themes/default/scripts/topic.js
#	Themes/default/webfonts/index.php

+++ Sources/Actions/Admin/RepairBoards.php
@@ -5,10 +5,10 @@
  *
  * @package SMF
  * @author Simple Machines https://www.simplemachines.org
- * @copyright 2025 Simple Machines and individual contributors
+ * @copyright 2026 Simple Machines and individual contributors
  * @license https://www.simplemachines.org/about/smf/license.php BSD
  *
- * @version 3.0 Alpha 3
+ * @version 3.0 Alpha 4
  */
 
 declare(strict_types=1);
@@ -1317,8 +1317,8 @@ class RepairBoards implements ActionInterface
 			],
 		);
 
-		$memberStartedID = (int) Board::getMsgMemberID($row['myid_first_msg']);
-		$memberUpdatedID = (int) Board::getMsgMemberID($row['myid_last_msg']);
+		$memberStartedID = (int) Board::getMsgMemberID((int) $row['myid_first_msg']);
+		$memberUpdatedID = (int) Board::getMsgMemberID((int) $row['myid_last_msg']);
 
 		$newTopicID = Db::$db->insert(
 			'',
